add chatbox - filter for admin order

+++ ElectronicCommerce/Areas/Admin/ViewModels/SaleReportOption.cs
@@ -5,6 +5,6 @@ namespace ElectronicCommerce.Areas.Admin.ViewModels
     public class SaleReportOption
     {
         public DateTime Time { get; set; }
-        public int TongTien { get; set; }
+        public decimal TongTien { get; set; }
     }
 }
